move replace redux to event
