+++ backend/routes/openEventRoutes.js
@@ -39,31 +39,4 @@ router.post('/join', async (req, res) => {
     );
 });
 
-/**
- * replaces the current guestsRSVP with the given rsvp
- * creates list of new users with name
- */
-router.post('/replace', async (req, res) => {
-  const event = await Event.findById(req.body.id);
-  event.guestsRSVP = [];
-  if (event) {
-    // split based on end of line
-    let rsvps = req.body.rsvps.split(/\r?\n/);
-    for (let ind in rsvps) {
-      const guest = await User({
-        name: rsvps[ind]
-      }).save();
-      event.guestsRSVP.push(guest);
-    }
-    event.save();
-    res.send(event);
-  } else {
-    console.error(
-      '[ERR] Event was not found. Passed in id: ' +
-        req.body.id +
-        ' in /event/replace'
-    );
-  }
-});
-
 module.exports = router;
