+++ frontend/actions/eventActions.js
@@ -4,6 +4,7 @@ import {
   GOT_ATTENDS,
   GOT_RSVPS,
   REPLACE_RSVPS,
+  REPLACED_RSVPS,
   RESET_EVENT,
   SELECT_EVENT
 } from './types';
@@ -30,6 +31,15 @@ export const getAttends = event => async dispatch => {
     qs.stringify({ id: event._id })
   );
   dispatch({ type: GOT_ATTENDS, payload: res.data });
+};
+
+export const replaceRsvps = (event, rsvps) => async dispatch => {
+  dispatch({ type: REPLACE_RSVPS });
+  const res = await axios.post(
+    '/api/event/replace',
+    qs.stringify({ id: event._id, rsvps: rsvps })
+  );
+  dispatch({ type: REPLACED_RSVPS, payload: res.data });
 };
 
 export const resetEvent = () => dispatch => {
