+++ frontend/actions/openEventActions.js
@@ -24,12 +24,3 @@ export const openJoinEvent = (code, id) => async dispatch => {
   // TODO: checkCode successfully joined or not
   dispatch({ type: OPEN_JOIN_EVENT, payload: res.data });
 };
-
-export const replaceRsvps = (event, rsvps) => async dispatch => {
-  dispatch({ type: REPLACE_RSVPS });
-  const res = await axios.post(
-    '/api/open/replace',
-    qs.stringify({ id: event._id, rsvps: rsvps })
-  );
-  dispatch({ type: REPLACED_RSVPS, payload: res.data });
-};
